plants-rocksdb: split the benchmark into modules
Break the single main.rs into focused modules:
- dataset.rs      dataset consts, Plant, CSV parsing, loading, id sampling
- bench.rs        shared plumbing: schema, RocksDB Db construction, by-id read,
                  EbsDelay, Metrics
- raw.rs          topology 1 (raw RocksDB)
- jazz.rs         topology 2 (Jazz + RocksDB)
- server.rs       topology 3 (client -> Jazz Server) + drain
- ws_transport.rs the WebSocket WireTransport
- main.rs         CLI parsing, dispatch, help

No behaviour change; output is byte-for-byte identical.
